requirements.txt
alembic==1.6.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.4.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==1.1.3
fastapi==0.68.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.14.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.13.4